drivers: led_strip: ws2812: Split the dts definition for LED and controller
The dts definition of rpi_pico-pio is divided so that ws2812 can be
described with a common dts definition, and the implementation
also follows.

Initially, there was some discussion, but with the introduction of #68514,
the situation has changed to the point where it is more desirable to
be able to reuse the ws2812 DTS.
