+++ drivers/led_strip/Kconfig
@@ -35,4 +35,6 @@ source "drivers/led_strip/Kconfig.apa102"
 
 source "drivers/led_strip/Kconfig.tlc5971"
 
+source "drivers/led_strip/Kconfig.rpi_pico"
+
 endif # LED_STRIP
