+++ drivers/led_strip/Kconfig.ws2812
@@ -49,7 +49,7 @@ config WS2812_STRIP_GPIO
 config WS2812_STRIP_RPI_PICO_PIO
 	bool "Raspberry Pi Pico PIO"
 	depends on DT_HAS_WORLDSEMI_WS2812_RPI_PICO_PIO_ENABLED
-	select PICOSDK_USE_PIO
+	depends on RASPBERRYPI_PICO_WS2812_CONTROLLER_PIO
 	help
 	  Use the PIO feature available on RaspberryPi Pico devices.
